CSSTUDIO-3176: Move files from /beta and delete /beta
