+++ jsconfig.json
@@ -7,7 +7,6 @@
       "components/*": ["src/components/*"],
       "views/*": ["src/views/*"],
       "hooks/*": ["src/hooks/*"],
-      "beta/*": ["src/beta/*"],
       "providers/*": ["src/providers/*"],
       "features/*": ["src/features/*"],
       "api/*": ["src/api/*"],
