--- src/beta/components/common/UserAvatar.jsx
@@ -1,37 +1,0 @@
-import { Avatar } from "@mui/material";
-/*
- * Demonstration by MUI of transforming string into a unique hex value
- */
-const stringToColor = (string) => {
-  let hash = 0;
-  let i;
-
-  /* eslint-disable no-bitwise */
-  for (i = 0; i < string.length; i += 1) {
-    hash = string.charCodeAt(i) + ((hash << 5) - hash);
-  }
-
-  let color = "#";
-
-  for (i = 0; i < 3; i += 1) {
-    const value = (hash >> (i * 8)) & 0xff;
-    color += `00${value.toString(16)}`.slice(-2);
-  }
-  /* eslint-enable no-bitwise */
-
-  return color;
-};
-
-const UserAvatar = ({ user, size = 36, ...props }) => {
-  return (
-    <Avatar
-      alt={user.userName}
-      sx={{ width: size, height: size, bgcolor: stringToColor(user.userName) }}
-      {...props}
-    >
-      {user.userName.slice(0, 2)}
-    </Avatar>
-  );
-};
-
-export default UserAvatar;
