--- src/beta/components/navigation/AppNavBar/UserMenu.jsx
@@ -1,19 +1,0 @@
-import { Box, Button } from "@mui/material";
-import { useShowLogout } from "features/authSlice";
-const UserMenu = ({ user }) => {
-  const { setShowLogout } = useShowLogout();
-
-  return (
-    <Box>
-      <Button
-        onClick={() => setShowLogout(true)}
-        variant="outlined"
-        color="primary"
-      >
-        {user.userName}
-      </Button>
-    </Box>
-  );
-};
-
-export default UserMenu;
