--- src/beta/components/navigation/AppNavBar/index.js
@@ -1,5 +1,0 @@
-import AppNavBar from "./AppNavBar";
-import UserMenu from "./UserMenu";
-
-export { AppNavBar, UserMenu };
-export default AppNavBar;
