+++ src/components/AppNavBar/index.js
@@ -1,0 +1,4 @@
+import AppNavBar from "./AppNavBar";
+
+export { AppNavBar };
+export default AppNavBar;
