+++ src/components/log/EntryTypeChip.jsx
