+++ src/components/log/LogDetails/AttachmentsGallery.jsx
