+++ src/components/log/LogDetails/CopyUrlButton.jsx
