+++ src/components/log/LogDetails/LogDetails.jsx
