+++ src/components/log/LogDetails/LogDetailsContainer.jsx
