+++ src/components/log/LogDetails/LogHeader.jsx
