+++ src/components/log/LogDetails/LogProperty.jsx
