+++ src/components/log/TextChip.jsx
